+++ src/Mutex.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Yiisoft\Mutex;
 
-use RuntimeException;
+use Yiisoft\Mutex\Exception\MutexReleaseException;
 
 use function md5;
 
@@ -54,7 +54,7 @@ abstract class Mutex implements MutexInterface
         }
 
         if (!$this->releaseLock()) {
-            throw new RuntimeException("Unable to release lock \"$this->mutexName\".");
+            throw new MutexReleaseException("Unable to release the \"$this->mutexName\" mutex.");
         }
 
         unset(self::$currentProcessLocks[$this->lockName]);