+++ src/MutexFactory.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Yiisoft\Mutex;
 
-use RuntimeException;
+use Yiisoft\Mutex\Exception\MutexLockedException;
 
 /**
  * Creates a mutex instance.
@@ -18,7 +18,7 @@ abstract class MutexFactory implements MutexFactoryInterface
         $mutex = $this->create($name);
 
         if (!$mutex->acquire($timeout)) {
-            throw new RuntimeException("Unable to acquire mutex \"$name\".");
+            throw new MutexLockedException("Unable to acquire the \"$name\" mutex.");
         }
 
         return $mutex;