+++ src/Synchronizer.php
@@ -5,8 +5,8 @@ declare(strict_types=1);
 namespace Yiisoft\Mutex;
 
 use ErrorException;
-use RuntimeException;
 use Throwable;
+use Yiisoft\Mutex\Exception\MutexLockedException;
 
 use function error_reporting;
 use function restore_error_handler;
@@ -16,7 +16,7 @@ use function set_error_handler;
  * Executes a callback in synchronized mode, i.e. only a single instance of the callback is executed at the same time:
  *
  * ```php
- * $synchronizer = new Synchronizer(new MyMutexFactory());
+ * $synchronizer = new \Yiisoft\Mutex\Synchronizer(new MyMutexFactory());
  *
  * $newCount = $synchronizer->execute('critical_logic', function () {
  *     return $counter->increase();
@@ -40,7 +40,7 @@ final class Synchronizer
      * @param int $timeout Time (in seconds) to wait for lock to be released. Defaults to zero meaning that
      * method {@see MutexInterface::acquire()} will return false immediately in case lock was already acquired.
      *
-     * @throws RuntimeException If unable to acquire lock.
+     * @throws MutexLockedException If unable to acquire lock.
      * @throws Throwable If an error occurred during the execution of the PHP callable.
      *
      * @return mixed The result of the PHP callable execution.