+++ tests/MutexFactoryTest.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace Yiisoft\Mutex\Tests;
 
 use PHPUnit\Framework\TestCase;
-use RuntimeException;
+use Yiisoft\Mutex\Exception\MutexLockedException;
 use Yiisoft\Mutex\MutexInterface;
 use Yiisoft\Mutex\Tests\Mocks\Mutex;
 use Yiisoft\Mutex\Tests\Mocks\MutexAcquired;
@@ -31,8 +31,8 @@ final class MutexFactoryTest extends TestCase
     {
         $factory = new MutexFactory(MutexAcquired::class);
 
-        $this->expectException(RuntimeException::class);
-        $this->expectExceptionMessage('Unable to acquire mutex "testCreateAndAcquireFailure".');
+        $this->expectException(MutexLockedException::class);
+        $this->expectExceptionMessage('Unable to acquire the "testCreateAndAcquireFailure" mutex.');
 
         $factory->createAndAcquire('testCreateAndAcquireFailure');
     }