+++ CHANGELOG.md
@@ -1,9 +1,8 @@
 # Yii Mutex Change Log
 
+## 1.1.0 under development
 
-## 1.0.2 under development
-
-- no changes in this release.
+- New #45: Add `MutexLockedException` and `MutexReleaseException` and throw them instead of `RuntimeException` (devanych)
 
 ## 1.0.1 August 19, 2021